feat: add TanStack code
react query added to manage querys and summarize the code `
