+++ src/main.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import { App } from './App.tsx'
+// import { App } from './AppOld.tsx'
+import { App } from './App'
 import './index.css'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
